JWT Authentication before deleting the post, JWT Authorization for deleting the post

+++ controllers/api/v1/posts_api_controller.js
@@ -12,12 +12,13 @@ module.exports.index = async function (req, res) {
 		let posts = await Post.find({})
 			// .sort({ createdAt: "desc" })
 			.sort("-createdAt") // To sort the posts in descending order.
-			.populate("user")
+			.populate("user", { password: 0 })
 			.populate({
 				path: "comments",
 				options: { sort: { createdAt: "desc" } },
 				populate: {
 					path: "user",
+					select: { password: 0 },
 				},
 			});
 		return res.status(200).json({
@@ -34,33 +35,46 @@ module.exports.destroy = async (req, res) => {
 	try {
 		let post = await Post.findById(req.params.id);
 
-		//If Post Image already exists in the Database
-		if (post.contentImage) {
-			//If Post Image already exists in the "/uploads/users/posts" Directory
-			if (fs.existsSync(path.join(__dirname, "..", post.contentImage))) {
-				//Delete that Post Image from the Directory
-				fs.unlinkSync(path.join(__dirname, "..", post.contentImage));
+		//User Authorization Check.
+		if (post.user == req.user.id) {
+			try {
+				console.log("Post User ", post.user, typeof post.user);
+				console.log("Req User ID ", req.user.id, typeof req.user.id);
+				console.log("Req User _ID ", req.user._id, typeof req.user._id);
+			} catch {}
+
+			//If Post Image already exists in the Database
+			if (post.contentImage) {
+				//If Post Image already exists in the "/uploads/users/posts" Directory
+				if (fs.existsSync(path.join(__dirname, "..", post.contentImage))) {
+					//Delete that Post Image from the Directory
+					fs.unlinkSync(path.join(__dirname, "..", post.contentImage));
+				}
 			}
-		}
 
-		//If Post Video already exists in the Database
-		if (post.contentVideo) {
-			//If Post Video already exists in the "/uploads/users/posts" Directory
-			if (fs.existsSync(path.join(__dirname, "..", post.contentVideo))) {
-				//Delete that Post Video from the Directory
-				fs.unlinkSync(path.join(__dirname, "..", post.contentVideo));
+			//If Post Video already exists in the Database
+			if (post.contentVideo) {
+				//If Post Video already exists in the "/uploads/users/posts" Directory
+				if (fs.existsSync(path.join(__dirname, "..", post.contentVideo))) {
+					//Delete that Post Video from the Directory
+					fs.unlinkSync(path.join(__dirname, "..", post.contentVideo));
+				}
 			}
+
+			//Delete the Post
+			post.remove();
+			//Delete the Post Comments
+			await Comment.deleteMany({ post: req.params.id });
+
+			//Return the Success Response
+			return res.status(200).json({
+				message: "Post & associated Comments Deleted Successfully",
+			});
+		} else {
+			return res.status(401).json({
+				message: "You are not Authorized to Delete this Post !!!",
+			});
 		}
-
-		//Delete the Post
-		post.remove();
-		//Delete the Post Comments
-		await Comment.deleteMany({ post: req.params.id });
-
-		//Return the Success Response
-		return res.status(200).json({
-			message: "Post & associated Comments Deleted Successfully",
-		});
 	} catch (err) {
 		//Return the Error Response
 		return res.status(500).json({
